extract base class for FragmentActivity
